Move files
Move HypercubeProtobufSerializer to package serialization
delete ProtobufObservationsSerializer
