+++ transmart-rest-api/src/main/groovy/org/transmartproject/rest/serialization/HypercubeProtobufSerializer.groovy
@@ -1,4 +1,4 @@
-package org.transmartproject.rest.protobuf
+package org.transmartproject.rest.serialization
 
 import com.google.common.collect.PeekingIterator
 import com.google.protobuf.Message
